requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.4.1
Django==2.0.13
django-cors-headers==2.4.0
django-filter==2.0.0
django-guardian==1.4.1
django-templated-mail==1.1.1
django-userena==2.0.1
djangorestframework==3.9.4
djoser==1.2.2
easy-thumbnails==2.3
fish==1
gunicorn==19.7.1
html2text==2014.12.29
idna==2.7
isodate==0.5.4
itypes==1.1.0
jdatetime==1.8.1
Jinja2==2.10
lxml==3.6.4
MarkupSafe==1.0
mysqlclient==1.3.13
panda==0.3.1
Pillow==5.2.0
PySimpleSOAP==1.16
python-docx==0.8.10
pytz==2019.3
requests==2.19.1
six==1.10.0
soap2py==1.16
uritemplate==3.0.0
urllib3==1.23
xlwt==1.1.2
zeep==0.18.1
